Fixes fsck issue
If we set atboot to true, it will write auto mount info to /etc/fstab.
When server restart, fsck.xfs will be called, it will check and repair
XFS filesystem in /etc/fstab. This will be inconsistent in /etc/fstab when we
removing an OSD. In this case, fsck.xfs will hang in server startup.

- add mount_at_boot parameter, we can set it false to avoid the inconsistent.
- add rspec test for the parameter