device.pp: extend to work without partition and reduce needed runs
Changed ceph::osd::device to work also without partition the device
and adding a GPT table to the device. This way we can use e.g. /dev/sdc
directly since there is no need to use a partition in most cases.

Reduce the needed puppet agent runs from at least 3  to max. 2 runs
if the MONs are running. Get the blkid if not available in facter
directly from the system and pass the information to a file under
/var/lib/ceph/tmp/. Then create the OSD without waiting for facter
by getting the information needed from this file. This way we can
spare one puppet run.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>