ceph::pool: fix conditions of exec calls to increase PGs/PGPs
Use onlyif and check also if the value of the current PGs/PGPs is the
same as the value we want to set.

Increase the time for the PGP increase command to 60 seconds and the
check interval from to every 5 seconds.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>
(cherry picked from commit cf1213ced42fba581b07117386295d925808eb0c)