conf.pp: add noexec to default mount options
By default prevent exectution of binaries on the OSD disks.
Usually there is no need to exec anything on these disks and
this also fix issues with hardening by preventing the scripts
to run find on OSD disks and subsequently filling up memory
and caches due to high number of files on these disks.

Fixes: SMBCAP-4597

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>