Reduce recovery priority
Default values for recovery can be a bit high depending on the CPU you
use. Thus adding new variable and using lower default help us to reduce
the impact on the CPUs. Sometimes this could led to crashes where OSDs
are exhausting all the CPU power and start committing suicide.

Signed-off-by: Sébastien Han <sebastien.han@enovance.com>