Fixing admin keyring exported resources.
When used in a 3 nodes configuration the exported key doesn't overwrite
the one already generated on the node. Because the exported key is the
only one that is added to ceph.conf, you cannot connect to ceph cluster
from the others nodes.
Also the admin key should never be empty since when you do that
ceph-authtool just create an empty file.